project/requirements.txt
Django==5.0.6
djangorestframework==3.15
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1